requirements.txt
Django
mysqlclient
celery
redis
djangorestframework
djangorestframework-simplejwt
